cloud/pyproject.toml
[tool.poetry]
name = "run"
version = "0.1.0"
description = "BQMS Cloud Run Image"
authors = ["A.J. Welch <ajwelch@google.com>"]

[tool.poetry.scripts]
bqms-cloud-run = 'run.main:main'

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-storage = "^2.5.0"
google-cloud-bigquery-migration = "^0.7.0"
PyYAML = "^6.0"
marshmallow = "^3.17.1"
cloudpathlib = "^0.10.0"

[tool.poetry.dev-dependencies]
black = "22.6.0"
isort = "^5.10.1"
pylint = "^2.15.0"
mypy = "^0.971"
nox = "^2022.8.7"
types-PyYAML = "^6.0.11"
types-protobuf = "^3.20.1"
types-requests = "^2.28.9"
types-setuptools = "^65.3.0"
yamllint = "^1.27.1"
coverage = "^6.4.4"
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# ########################
# ##### BLACK
# ########################

# [Docs root]
#   https://black.readthedocs.io/en/stable/
# [Config option reference]
#   https://black.readthedocs.io/en/stable/usage_and_configuration/the_basics.html#command-line-options

[tool.black]
# Black will refuse to run if it's not this version.
required-version = "22.6.0"

# Ensure black's output will be compatible with all listed versions.
target-version = ["py310"]

# ########################
# ##### ISORT
# ########################

# [Docs root]
#   https://pycqa.github.io/isort/
# [Config option reference]
#   https://pycqa.github.io/isort/docs/configuration/options.html

# NOTE: File inclusion//exclusion/skip options are set at the invocation site
# and shouldn't be set in this config file.

[tool.isort]

# Ensures isort classifies imports from `run` as first-party.
known_first_party = ["run"]

# Sets a variety of default options for parentheses etc that are compatible with black.
profile = "black"

# Sorts uppercase imports before lowercase improts.
case_sensitive=true

# ########################
# ##### PYLINT
# ########################

# [Docs root]
#   https://pylint.pycqa.org/en/latest/
# [Rule/options reference]
#   https://pylint.pycqa.org/en/latest/technical_reference/features.html
#   This is a comprehensive reference of all Pylint checkers/rules and seems to be one of the only
#   docs pages that's actually maintained.
# [Changelog]
#   https://pylint.pycqa.org/en/latest/whatsnew/index.html
#   Often contains up-to-date info that corrects incorrect stuff in the older docs.
# [Github issues]
#   https://github.com/PyCQA/pylint/issues
#   Best places to find workarounds or discussion/explanation of weird behavior.

# Pylint has many options that are divvied up by category. Each [SECTION] below corresponds to a
# section in the "Rule/options reference" doc.

[tool.pylint.master]

# Minimum python version.
py-version = "3.10"

[tool.pylint.messages-control]
disable=[
    # [checker] redundant with mypy
    "typecheck",
    # Sometimes it's useful to leave `TODO` comments in the code.
    "fixme",
]

[tool.pylint.format]

# https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html#pylint
max-line-length = "88"

[tool.pylint.basic]

# Minimum line length for functions/classes that require docstrings, shorter ones are exempt.
docstring-min-length = "10"

[tool.pylint.design]

# Maximum number of attributes for a class.
max-attributes="8"

# Maximum number of arguments for function / method.
max-args = "8"

# ########################
# ##### MYPY
# ########################

# [Docs root]
#   https://mypy.readthedocs.io/en/stable/index.html
# [Config option reference]
#   https://mypy.readthedocs.io/en/stable/config_file.html#the-mypy-configuration-file

[tool.mypy]

# google.cloud.storage missing library stubs or py.typed marker
ignore_missing_imports = true

# If a type stub package is missing from the environment but available on typeshed, automatically
# install it. Note that without `--non-interactive` this will prompt the user.
install_types = true

# Allow PEP-420-style namespace packages. Without this enabled, different parts of a namespace
# package will trigger "Duplicate module" errors.
namespace_packages = true

# Print codes (e.g. "[arg-type]") for each error in output.
show_error_codes = true

# ########################
# ##### PYTEST
# ########################

# [Docs root]
#   https://docs.pytest.org/en/stable/
# [Config option reference]
#   https://docs.pytest.org/en/stable/reference/customize.html

[tool.pytest.ini_options]

# https://docs.pytest.org/en/stable/how-to/capture-warnings.html#controlling-warnings
filterwarnings = [
    # https://docs.pytest.org/en/stable/how-to/capture-warnings.html#deprecationwarning-and-pendingdeprecationwarning
    "ignore:distutils Version classes are deprecated. Use packaging.version instead.:DeprecationWarning:marshmallow",
]
